+++ js/jquery.terminal.js
@@ -41,7 +41,7 @@
  *
  * broken image by Sophia Bai from the Noun Project (CC-BY)
  *
- * Date: Sun, 23 Jun 2024 13:34:27 +0000
+ * Date: Sun, 07 Jul 2024 21:51:53 +0000
  */
 /* global define, Map, BigInt */
 /* eslint-disable */
@@ -5311,7 +5311,7 @@
     // -------------------------------------------------------------------------
     $.terminal = {
         version: 'DEV',
-        date: 'Sun, 23 Jun 2024 13:34:27 +0000',
+        date: 'Sun, 07 Jul 2024 21:51:53 +0000',
         // colors from https://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',
@@ -10991,18 +10991,29 @@
                             unpromise(next, function() {
                                 // extended commands should be processed only
                                 // once in echo and not on redraw
+                                var continuation_run = locals.flush && self.find('.partial').length;
                                 if (locals.flush) {
+                                    self.stopTime('echo');
                                     self.flush();
                                     // HACK: when rendering multiple partials
                                     // we need to force and update them all at once
                                     // so we have line breaks #952
                                     if (self.find('.partial').length) {
-                                        var last_line = lines.last_line();
-                                        //self.update(-1, last_line[0], last_line[1]);
+                                        // setTimeout is required when you echo async function
+                                        // and immediately after echo someting else
+                                        self.oneTime(1, 'echo', function() {
+                                            var last_line = lines.last_line();
+                                            if (last_line) {
+                                                self.update(-1, last_line[0], last_line[1]);
+                                            }
+                                            cont();
+                                        });
                                     }
                                     fire_event('onAfterEcho', [arg]);
                                 }
-                                cont();
+                                if (!continuation_run) {
+                                    cont();
+                                }
                             }, error);
                         }, error);
                     } catch (e) {