+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * fix line wrapping when calling multiple echo `newline:false` [#952](https://github.com/jcubic/jquery.terminal/issues/952)
 * fix cursor position when echo `newline:false` fill the line
 * fix running typing animation on empty terminal [#950](https://github.com/jcubic/jquery.terminal/issues/950)
+* get rid of canvas warning from images in less [#955](https://github.com/jcubic/jquery.terminal/issues/955)
 
 ## 2.42.0
 ### Features