+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * fix cursor position when echo `newline:false` fill the line
 * fix running typing animation on empty terminal [#950](https://github.com/jcubic/jquery.terminal/issues/950)
 * get rid of canvas warning from images in less [#955](https://github.com/jcubic/jquery.terminal/issues/955)
+* fix update with newline after echo without it [#954](https://github.com/jcubic/jquery.terminal/issues/954)
 
 ## 2.42.0
 ### Features